Use item-based not region-based trawling.
Essentially, for my usecase it's more helpful to have a global picture
of prices for an item than to have a complete regional view of all
items. Here, we use a priority queue to get the item updated longest
ago, poll all regions for it, then put it back into the queue.