Don't split orders for the same type over different messages.
Many clients use an implicit assumption that all orders for a given
(typeID, regionID) are in a single Rowset in an EMDR message. Previously
that was fine since cache scrapers worked on a per-region-per-item
basis.

The CREST bulk endpoint allows all market orders for a region to be
downloaded. EMDR's HTTP upload has a limited size, so we were splitting
the data into chunks of at most 30,000 orders - but breaking that
implicit assumption.

Now, we are more clever, and keeping groups of orders for a single type
together while still breaking into chunks of no more than 30,000 orders
at once.