Even longer timeout.
The usual response time is ~30s, but can be much larger - especially if the
trawler is currently processing The Forge. This is because we try to delete
expired orders before querying for prices.

Perhaps the real solution is to not do that, but include a WHERE in the view
used to filter them out? We'd still want to tidy up periodically but this could
be done after the fact.